requirements.txt
fastapi[standard]
websockets
sqlmodel
python-dotenv
uvicorn
pymysql
passlib